provide dynamic radio button for new project
